+++ api/prisma/migrations/20260120195406_cron_scheduler_market_requests/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE `symbols` MODIFY `marketSyncStatus` VARCHAR(191) NULL;